refactor(ui): production residual sweep to size limits
