+++ src/quodeq/ui/src/appShellProps.js
@@ -1,4 +1,5 @@
 import { shouldShowEvaluateButton, shouldShowProjectTabs, shouldShowCompareTab } from './appGating.js';
+import { buildDashboardDataBundle, buildNavigationBundle } from './routes/renderers.jsx';
 
 // Pure prop-builders for App.jsx's Sidebar/TopBar wiring, extracted
 // verbatim from the inline JSX props so App.jsx itself stays a thin
@@ -53,5 +54,43 @@ export function buildTopBarProps({
     onBack: navPop,
     effectiveDark,
     onToggleTheme,
+  };
+}
+
+// The MainContent/route props bundle — extracted verbatim from App.jsx's
+// inline object literal. Pure: every field is either passed straight
+// through or built from an existing builder (buildDashboardDataBundle /
+// buildNavigationBundle), no new derivation.
+export function buildContentProps({
+  state, sharedSignal, navTab, navStackLength, isEvaluating, showToast, setWizardEntry,
+  dismissFinding, applyDelta, bumpDismissRefresh, dismissRefreshKey,
+}) {
+  return {
+    dashboardData: buildDashboardDataBundle({ state, sharedHasContent: sharedSignal.hasContent }),
+    navigation: buildNavigationBundle({
+      state, navTab, navStackLength,
+      isEvaluating, showToast, setWizardEntry,
+      sharedHasContent: sharedSignal.hasContent,
+    }),
+    evaluation: state.evalLifecycle,
+    serverHealth: { connected: state.serverConnected, setConnected: state.setServerConnected },
+    settings: state.settings,
+    refreshDashboard: state.refreshDashboard,
+    // Debounced ACTIVE reconcile for suppression mutations (dismiss/restore/
+    // delete) — see useDashboard.js. refreshDashboard's refetchType:'none'
+    // only marks the cache stale; this actually refetches the always-mounted
+    // Overview observer after the 1200ms window, so restore-all/delete-all
+    // (whose response can't be patched via applyMutationDelta) and every
+    // other suppression mutation converge without waiting for a project
+    // switch.
+    scheduleDashboardReconcile: state.scheduleDashboardReconcile,
+    dismissFinding,
+    // Patch the dashboard/scores caches from the dismiss response delta so the
+    // Overview updates instantly. Additive — the refreshDashboard /
+    // bumpDismissRefresh mechanisms below still run. The delta carries only the
+    // mutation shape; the caller folds in the rescored dims from result.scores.
+    applyDelta,
+    bumpDismissRefresh,
+    dismissRefreshKey,
   };
 }
