+++ src/quodeq/ui/src/features/dashboard/hooks/useDashboardHandlers.js
@@ -1,0 +1,15 @@
+import { useMemo } from 'react';
+
+export function useDashboardHandlers(onNavigate, dashboard) {
+  return useMemo(() => ({
+    handleDimensionCardClick: (item, runId) => {
+      if (!onNavigate) return;
+      const dateLabel = dashboard?.selectedRun?.dateLabel || item.fromDateLabel;
+      onNavigate('explorer', { dimension: item.dimension, runId: runId || item.fromRunId, dateLabel, fromProject: item.fromProject });
+    },
+    handleAccumulatedDimensionClick: (item) => {
+      if (onNavigate) onNavigate('explorer', { dimension: item.dimension, runId: item.fromRunId, dateLabel: item.fromDateLabel, fromProject: item.fromProject });
+    },
+    handleFileClick: (fileObj) => { if (onNavigate) onNavigate('file', { file: fileObj }); },
+  }), [onNavigate, dashboard]);
+}
